deployer: restructure code
