+++ cmd/deploy/main.go
@@ -1,8 +1,10 @@
 package main
 
 import (
+	"context"
 	"os"
 
+	"github.com/golang/protobuf/jsonpb"
 	"github.com/navikt/deployment/pkg/deployer"
 	"github.com/navikt/deployment/pkg/pb"
 
@@ -11,29 +13,59 @@ import (
 )
 
 func main() {
-	cfg := deployer.NewConfig()
+	err := run()
+	if err == nil {
+		return
+	}
+	code := deployer.ErrorExitCode(err)
+	if code == deployer.ExitInvocationFailure {
+		flag.Usage()
+	}
+	log.Errorf("fatal: %s", err)
+	os.Exit(int(code))
+}
 
+func run() error {
+	// Configuration and context
+	cfg := deployer.NewConfig()
+	ctx, cancel := context.WithTimeout(context.Background(), cfg.Timeout)
+	defer cancel()
+
+	// Logging
 	deployer.SetupLogging(*cfg)
 
+	// Set up asynchronous gRPC connection
 	grpcConnection, err := deployer.NewGrpcConnection(*cfg)
 	if err != nil {
-		log.Errorf("fatal: %s", err)
-		os.Exit(int(deployer.ExitUnavailable))
+		return err
 	}
-	defer grpcConnection.Close()
-
-	grpcClient := pb.NewDeployClient(grpcConnection)
-
-	d := deployer.Deployer{Client: grpcClient}
-
-	code, err := d.Run(*cfg)
-
-	if err != nil {
-		if code == deployer.ExitInvocationFailure {
-			flag.Usage()
+	defer func() {
+		err := grpcConnection.Close()
+		if err != nil {
+			log.Error(err)
 		}
-		log.Errorf("fatal: %s", err)
+	}()
+
+	d := deployer.Deployer{
+		Client: pb.NewDeployClient(grpcConnection),
 	}
 
-	os.Exit(int(code))
+	request, err := d.Prepare(ctx, *cfg)
+	if err != nil {
+		return err
+	}
+
+	if cfg.PrintPayload {
+		marsh := jsonpb.Marshaler{Indent: "  "}
+		err = marsh.Marshal(os.Stdout, request)
+		if err != nil {
+			log.Errorf("print payload: %s", err)
+		}
+	}
+
+	if cfg.DryRun {
+		return nil
+	}
+
+	return d.Deploy(ctx, request)
 }
